LayoutTests/editing/selection/caret-in-empty-block-vertical-rl-rtl-expected.txt

This test verifies that an empty block contenteditable element with a vertical writing mode, and right-to-left text direction, gets a valid caret rect.


PASS caretRect.left is correct.
PASS caretRect.top is correct.
PASS caretRect.width is correct.
PASS caretRect.height is correct.
PASS successfullyParsed is true

TEST COMPLETE
